Added HashFunction classes
There could be multiple ways to implement hash functions,
so all have got to implement the HashFunction interface,
to be usable by the BloomFilter classes.

The easiest one could be a splitting of the given hash digit.
The next one could be a simple salted hash function.
At last, there could be flexible hash function containing
a group of selfgenerated functions.