Added Set methods and Tests
The Set class implements now the most algorithms in a generic way,
needed by subclasses. All methods of the Set class now have got an
unittest to evaluate its methods.

The Trie Interface got a new method to erase an entry.

CountingBloomFilter are now needed for the Set to erase an entry.