Added toString method and changed bit pos calc
toString method added for better debugging

The calculated bit positions are changed to a more intuitive
way of calculation. Now, the index in the byte array will be
calculated, not the position in the bit array. As second index,
the position of the bit at the selected byte is calculated.