Added a UniqueFilterDiffHandler
The new class is a subclass of the C_DiffHandler and extends the
C_DiffHandler by adding a list of already handled hashes. So the
called will only be called once per hash.

This implementation should only be chosen, if the difference is small
and the list of hashes fits into the memory. Otherwise, the normal
C_DiffHandler should be used, because it is stateless.