Added diff method to DBTrie
AbstractDiffHandler is now extended to provide the information, which
side misses a handled hash.
DBTrie has now a diff method, which looks exactly like the BloomFilter
diff method. It takes a subtree cut through a Trie and checks, if
there are entries, which are not present in the actual Trie instance,
all missing entries are handled by the given AbstractDiffHandler
instance