Added new Message Types to Packets
A ERROR_MSG can be used to send a Failure Description
A TRIE_ROOT message can be send to proof, if the trie root is equal.
This should be done at the beginning of a sync session. It could also
be sent at the end, to ensure equal tries, if multiple sync sessions
are executed in parallel.